libdfx initial commit
Some systems are memory constrained but they need to load very
large Configuration files. The FPGA subsystem allows drivers to
request this Configuration image be loaded from the filesystem,
but this requires that the entire configuration data be loaded
into kernel memory first before it's provided to the driver.
This can lead to a situation where we map the configuration
data twice, once to load the configuration data into kernel
memory and once to copy the configuration data into the final
resting place which is nothing but a dma-able continuous buffer.

This creates needless memory pressure and delays due to multiple
copies. The libdfx library allows drivers to load the Configuration
data directly from a pre-allocated buffer.This skips the intermediate
step of allocating a buffer in kernel memory to hold the Configuration
data.

Signed-off-by: Nava kishore Manne <nava.manne@xilinx.com>