Src: Fix for improper overlay status check
Upstream overlay framework incorrectly updating the configfs
‘status’  to ‘applied’ when there is an error in create_overlay()
and libdfx checks the content of the status file to decide if the
overlay was a success or not. Hence Libdfx API’s doesn’t return
a proper status in certain use cases. The patch provides a workaround
by updating the overlay status check to path file instead of getting
it from status file.

Signed-off-by: Nava kishore Manne <nava.manne@xilinx.com>