libdfx: src: Make word align the configuration data
For ZynqMP platform to avoid the unwanted copies at firmware(PMUFW)
this patch provides a fix to align PL configuration data if the data
is not word-aligned. It improves the PL configuration timings to
process unaligned bit/bin files.

To align the configuration data this patch adds a few padding bytes and
these additional padding bytes will not create any functional impact on
the PL configuration.

Signed-off-by: Nava kishore Manne <nava.manne@xilinx.com>